Error better on failed open.  Comment.
If we can't open a file, say which index it is.  Check
in both places we open files, not just one.

Add comment pointing out that the current logic does not (I think)
handle the case of finding that the event is in an overlap region after
no events having been in the non-overlap region. i.e. the case where we
need to open both files on the same event.