Flip order of functions, use getopt.
Put main at the bottom where it is right and proper for it to be
and get rid of the prototypes.

Use getopt to get options, make options for changing the chunksize,
overlap size, maximum number of output files, to not use the database.
Also take the input file and out basename with options.
Add help text.