Set padding bytes to zero.
Previously, when writing out padding bytes to fill out
a steering block, uh, thing, whatever happened to be in
mbuf after the real content was blindly written to file.
Now memset everthing after the real information to zero
so that we don't confusingly write out duplicate events
(that are ignored later because of the record type).
This will also improve compression if these files ever
need to be compressed.